+++ test/Play.cpp
@@ -6,13 +6,13 @@
  * Notes:
  * Using Clang && GCC on 32-Bit:
  * sizeof (Piece) = 4
- * sizeof (size_t) = 4
- * sizeof (Point3) = 12
- * sizeof (IndexedPiece) = 16
+ * sizeof (uint8_t) = 1
+ * sizeof (Point3) = 3
+ * sizeof (IndexedPiece) = ?
  *
  * On 64-Bit:
  * sizeof (Piece) = TBD
- * sizeof (size_t) = TBD
+ * sizeof (uint8_t) = TBD
  * sizeof (Point3) = TBD
  * sizeof (IndexedPiece) = TBD
  */
@@ -27,6 +27,21 @@ char getTailChar(const Tail& tail);
 void print(const Piece& piece);
 void placePieces(Game& game);
 void displayBoard(const Board& board);
+void displayPlayerSet(const Player& player);
+void displayPieces(const Game& game);
+
+void displayState(const Game& game)
+{
+    const Board* board = game.gameBoard();
+    const Player* one = game.playerOne();
+    const Player* two = game.playerTwo();
+    displayBoard(*board);
+    cout << "Player One: " << endl;
+    displayPlayerSet(*one);
+    cout << "Player Two: " << endl;
+    displayPlayerSet(*two);
+//    displayPieces(game);
+}
 
 int main()
 {
@@ -41,9 +56,16 @@ int main()
     game.start(); // Set Phase to Placement
     placePieces(game);
     game.start(); // Set Phase to Running
-
-    const Board* board = &game.gameBoard();
-    displayBoard(*board);
+    displayState(game);
+    for (SizeType i = 0; i < 9; ++i)
+    {
+        SizeType soldier = 14 + i;
+        game.move(soldier, Move(1, Direction::N));
+        displayState(game);
+        game.move(soldier, Move(1, Direction::N));
+        displayState(game);
+    }
+    
     return 0;
 }
 
@@ -120,118 +142,104 @@ void print(const Piece& piece)
 
 void placePieces(Game& game)
 {
-    std::cout << game.drop(0,  SmallPoint3(0, 5, 2)) << std::endl; // Commander
-    std::cout << game.drop(0,  SmallPoint3(0, 5, 2)) << std::endl; // Commander
-    std::cout << game.drop(1,  SmallPoint3(1, 5, 2)) << std::endl; // Captain 
-    std::cout << game.drop(1,  SmallPoint3(1, 5, 2)) << std::endl; // Captain 
-    std::cout << game.drop(2,  SmallPoint3(0, 4, 2)) << std::endl; // Captain 
-    std::cout << game.drop(2,  SmallPoint3(0, 4, 2)) << std::endl; // Captain 
-    std::cout << game.drop(3,  SmallPoint3(1, 1, 0)) << std::endl; // Samurai
-    std::cout << game.drop(3,  SmallPoint3(1, 1, 0)) << std::endl; // Samurai 
-    std::cout << game.drop(4,  SmallPoint3(2, 0, 0)) << std::endl; // Samurai 
-    std::cout << game.drop(4,  SmallPoint3(2, 0, 0)) << std::endl; // Samurai 
-    std::cout << game.drop(5,  SmallPoint3(2, 6, 0)) << std::endl; // Ninja 
-    std::cout << game.drop(5,  SmallPoint3(2, 8, 0)) << std::endl; // Ninja 
-    std::cout << game.drop(6,  SmallPoint3(3, 9, 0)) << std::endl; // Ninja 
-    std::cout << game.drop(6,  SmallPoint3(3, 9, 0)) << std::endl; // Ninja 
-    std::cout << game.drop(7,  SmallPoint3(3, 5, 0)) << std::endl; // Ninja 
-    std::cout << game.drop(7,  SmallPoint3(9, 1, 3)) << std::endl; // Ninja 
-    std::cout << game.drop(8,  SmallPoint3(4, 0, 0)) << std::endl; // Catapult 
-    std::cout << game.drop(8,  SmallPoint3(4, 0, 0)) << std::endl; // Catapult 
-    std::cout << game.drop(9,  SmallPoint3(4, 1, 0)) << std::endl; // Fortress 
-    std::cout << game.drop(9,  SmallPoint3(4, 1, 0)) << std::endl; // Fortress 
-    std::cout << game.drop(10, SmallPoint3(5, 0, 0)) << std::endl; // Hidden Dragon 
-    std::cout << game.drop(10, SmallPoint3(5, 0, 0)) << std::endl; // Hidden Dragon 
-    std::cout << game.drop(11, SmallPoint3(5, 1, 0)) << std::endl; // Prodigy 
-    std::cout << game.drop(11, SmallPoint3(5, 1, 0)) << std::endl; // Prodigy 
-    std::cout << game.drop(12, SmallPoint3(6, 0, 0)) << std::endl; // Archer
-    std::cout << game.drop(12, SmallPoint3(6, 0, 0)) << std::endl; // Archer
-    std::cout << game.drop(13, SmallPoint3(6, 1, 0)) << std::endl; // Archer
-    std::cout << game.drop(13, SmallPoint3(6, 1, 0)) << std::endl; // Archer
-    std::cout << game.drop(14, SmallPoint3(0, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(14, SmallPoint3(0, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(15, SmallPoint3(1, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(15, SmallPoint3(1, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(16, SmallPoint3(2, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(16, SmallPoint3(2, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(17, SmallPoint3(3, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(17, SmallPoint3(3, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(18, SmallPoint3(4, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(18, SmallPoint3(4, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(19, SmallPoint3(5, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(19, SmallPoint3(5, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(20, SmallPoint3(6, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(20, SmallPoint3(6, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(21, SmallPoint3(8, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(21, SmallPoint3(8, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(22, SmallPoint3(8, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(22, SmallPoint3(8, 2, 0)) << std::endl; // Soldier 
+    game.drop(0,  SmallPoint3(0, 0, 0)); // Commander
+    game.drop(0,  SmallPoint3(0, 0, 0)); // Commander
+    game.drop(1,  SmallPoint3(1, 0, 0)); // Captain 
+    game.drop(1,  SmallPoint3(1, 0, 0)); // Captain 
+    game.drop(2,  SmallPoint3(0, 1, 0)); // Captain 
+    game.drop(2,  SmallPoint3(0, 1, 0)); // Captain 
+    game.drop(3,  SmallPoint3(1, 1, 0)); // Samurai
+    game.drop(3,  SmallPoint3(1, 1, 0)); // Samurai 
+    game.drop(4,  SmallPoint3(2, 0, 0)); // Samurai 
+    game.drop(4,  SmallPoint3(2, 0, 0)); // Samurai 
+    game.drop(5,  SmallPoint3(2, 1, 0)); // Ninja 
+    game.drop(5,  SmallPoint3(2, 1, 0)); // Ninja 
+    game.drop(6,  SmallPoint3(3, 0, 0)); // Ninja 
+    game.drop(6,  SmallPoint3(3, 0, 0)); // Ninja 
+    game.drop(7,  SmallPoint3(3, 1, 0)); // Ninja 
+    game.drop(7,  SmallPoint3(3, 1, 0)); // Ninja 
+    game.drop(8,  SmallPoint3(4, 0, 0)); // Catapult 
+    game.drop(8,  SmallPoint3(4, 0, 0)); // Catapult 
+    game.drop(9,  SmallPoint3(4, 1, 0)); // Fortress 
+    game.drop(9,  SmallPoint3(4, 1, 0)); // Fortress 
+    game.drop(10, SmallPoint3(5, 0, 0)); // Hidden Dragon 
+    game.drop(10, SmallPoint3(5, 0, 0)); // Hidden Dragon 
+    game.drop(11, SmallPoint3(5, 1, 0)); // Prodigy 
+    game.drop(11, SmallPoint3(5, 1, 0)); // Prodigy 
+    game.drop(12, SmallPoint3(6, 0, 0)); // Archer
+    game.drop(12, SmallPoint3(6, 0, 0)); // Archer
+    game.drop(13, SmallPoint3(6, 1, 0)); // Archer
+    game.drop(13, SmallPoint3(6, 1, 0)); // Archer
+    game.drop(14, SmallPoint3(0, 2, 0)); // Soldier 
+    game.drop(14, SmallPoint3(0, 2, 0)); // Soldier 
+    game.drop(15, SmallPoint3(1, 2, 0)); // Soldier 
+    game.drop(15, SmallPoint3(1, 2, 0)); // Soldier 
+    game.drop(16, SmallPoint3(2, 2, 0)); // Soldier 
+    game.drop(16, SmallPoint3(2, 2, 0)); // Soldier 
+    game.drop(17, SmallPoint3(3, 2, 0)); // Soldier 
+    game.drop(17, SmallPoint3(3, 2, 0)); // Soldier 
+    game.drop(18, SmallPoint3(4, 2, 0)); // Soldier 
+    game.drop(18, SmallPoint3(4, 2, 0)); // Soldier 
+    game.drop(19, SmallPoint3(5, 2, 0)); // Soldier 
+    game.drop(19, SmallPoint3(5, 2, 0)); // Soldier 
+    game.drop(20, SmallPoint3(6, 2, 0)); // Soldier 
+    game.drop(20, SmallPoint3(6, 2, 0)); // Soldier 
+    game.drop(21, SmallPoint3(7, 2, 0)); // Soldier 
+    game.drop(21, SmallPoint3(7, 2, 0)); // Soldier 
+    game.drop(22, SmallPoint3(8, 2, 0)); // Soldier 
+    game.drop(22, SmallPoint3(8, 2, 0)); // Soldier 
 }
-/*
-void placePieces(Game& game)
-{
-    std::cout << game.drop(0,  SmallPoint3(0, 0, 0)) << std::endl; // Commander
-    std::cout << game.drop(0,  SmallPoint3(0, 0, 0)) << std::endl; // Commander
-    std::cout << game.drop(1,  SmallPoint3(1, 0, 0)) << std::endl; // Captain 
-    std::cout << game.drop(1,  SmallPoint3(1, 0, 0)) << std::endl; // Captain 
-    std::cout << game.drop(2,  SmallPoint3(0, 1, 0)) << std::endl; // Captain 
-    std::cout << game.drop(2,  SmallPoint3(0, 1, 0)) << std::endl; // Captain 
-    std::cout << game.drop(3,  SmallPoint3(1, 1, 0)) << std::endl; // Samurai
-    std::cout << game.drop(3,  SmallPoint3(1, 1, 0)) << std::endl; // Samurai 
-    std::cout << game.drop(4,  SmallPoint3(2, 0, 0)) << std::endl; // Samurai 
-    std::cout << game.drop(4,  SmallPoint3(2, 0, 0)) << std::endl; // Samurai 
-    std::cout << game.drop(5,  SmallPoint3(2, 1, 0)) << std::endl; // Ninja 
-    std::cout << game.drop(5,  SmallPoint3(2, 1, 0)) << std::endl; // Ninja 
-    std::cout << game.drop(6,  SmallPoint3(3, 0, 0)) << std::endl; // Ninja 
-    std::cout << game.drop(6,  SmallPoint3(3, 0, 0)) << std::endl; // Ninja 
-    std::cout << game.drop(7,  SmallPoint3(3, 1, 0)) << std::endl; // Ninja 
-    std::cout << game.drop(7,  SmallPoint3(3, 1, 0)) << std::endl; // Ninja 
-    std::cout << game.drop(8,  SmallPoint3(4, 0, 0)) << std::endl; // Catapult 
-    std::cout << game.drop(8,  SmallPoint3(4, 0, 0)) << std::endl; // Catapult 
-    std::cout << game.drop(9,  SmallPoint3(4, 1, 0)) << std::endl; // Fortress 
-    std::cout << game.drop(9,  SmallPoint3(4, 1, 0)) << std::endl; // Fortress 
-    std::cout << game.drop(10, SmallPoint3(5, 0, 0)) << std::endl; // Hidden Dragon 
-    std::cout << game.drop(10, SmallPoint3(5, 0, 0)) << std::endl; // Hidden Dragon 
-    std::cout << game.drop(11, SmallPoint3(5, 1, 0)) << std::endl; // Prodigy 
-    std::cout << game.drop(11, SmallPoint3(5, 1, 0)) << std::endl; // Prodigy 
-    std::cout << game.drop(12, SmallPoint3(6, 0, 0)) << std::endl; // Archer
-    std::cout << game.drop(12, SmallPoint3(6, 0, 0)) << std::endl; // Archer
-    std::cout << game.drop(13, SmallPoint3(6, 1, 0)) << std::endl; // Archer
-    std::cout << game.drop(13, SmallPoint3(6, 1, 0)) << std::endl; // Archer
-    std::cout << game.drop(14, SmallPoint3(0, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(14, SmallPoint3(0, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(15, SmallPoint3(1, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(15, SmallPoint3(1, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(16, SmallPoint3(2, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(16, SmallPoint3(2, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(17, SmallPoint3(3, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(17, SmallPoint3(3, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(18, SmallPoint3(4, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(18, SmallPoint3(4, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(19, SmallPoint3(5, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(19, SmallPoint3(5, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(20, SmallPoint3(6, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(20, SmallPoint3(6, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(21, SmallPoint3(7, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(21, SmallPoint3(7, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(22, SmallPoint3(8, 2, 0)) << std::endl; // Soldier 
-    std::cout << game.drop(22, SmallPoint3(8, 2, 0)) << std::endl; // Soldier 
-}
-*/
+
 void displayBoard(const Board& board)
 {
-    /*
-    for (uint8_t ht = 0; ht < board.getHeight(); ++ht)
+    for (int j = 8; j != -1; --j)
     {
-        for (uint8_t wd = 0; wd < board.getWidth(); ++wd)
+        for (SizeType k = 0; k < board.getHeight(); ++k)
         {
-            for (uint8_t dp = 8; dp < board.getDepth(); --dp)
+            for (SizeType i = 0; i < board.getWidth(); ++i)
             {
-                auto p = board(wd, dp, ht);
-                print(*p);
+                cout << '|';
+                auto p = board(i,j,k);
+                print (*p);
             }
-            std::cout << "|";
+            cout << "|  ";
         }
+        cout << endl;
     }
-    */
+}
+
+void displayPlayerSet(const Player& player)
+{
+    auto pieceSet = player.getFullSet();
+    
+    for (SizeType i = 0; i < pieceSet.Set.size(); ++i)
+    {
+        cout << "Index: " << (size_t) i << " --> Piece: ";
+        const auto piece = pieceSet.pieceAt(i);
+        if (piece.onHead()) 
+            cout << getHeadString(piece) << " --> ";
+        else 
+            cout << getTailString(piece) << " --> ";
+        cout << " Point3: " << pieceSet.pointAt(i) << "  |  ";
+        if (i != 0 && i % 2 == 0) cout << endl;
+    }
+    cout << endl;
+}
+
+void displayPieces(const Game& game)
+{
+    auto pieceSet = game.currentPlayer().getFullSet();
+    
+    for (SizeType i = 0; i < pieceSet.Set.size(); ++i)
+    {
+        cout << "Index: " << (size_t) i << " --> Piece: ";
+        const auto piece = pieceSet.pieceAt(i);
+        if (piece.onHead()) 
+            cout << getHeadString(piece) << " --> ";
+        else 
+            cout << getTailString(piece) << " --> ";
+        cout << " Point3: " << pieceSet.pointAt(i) << "  |  ";
+        if (i != 0 && i % 2 == 0) cout << endl;
+    }
+    cout << endl;
 }